Changed build to Maven project. Upgraded to Java 11
