+++ .idea/libraries/simply.xml
@@ -1,0 +1,9 @@
+<component name="libraryTable">
+  <library name="simply">
+    <CLASSES>
+      <root url="jar://$PROJECT_DIR$/simply/dist/simply.jar!/" />
+    </CLASSES>
+    <JAVADOC />
+    <SOURCES />
+  </library>
+</component>
