+++ doc/simplyClassDiagram.png
