+++ doc/simplyClassDiagram.ucd
