+++ doc/simplyClassDiagramUMLStyle.png
