+++ doc/simplyStateDiagram.png
