+++ doc/simplyStateDiagram.utd
