--- simply/.settings/oracle.eclipse.tools.webtier.ui.prefs
@@ -1,3 +1,0 @@
-#Mon Sep 22 12:51:47 CEST 2014
-_hiddenCategory_\:HTML=false
-eclipse.preferences.version=1
