--- simply/.umlproject
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="ASCII"?>
-<properties:ProjectDescription xmi:version="2.0" xmlns:xmi="http://www.omg.org/XMI" xmlns:properties="properties.xmi" modelDir="model" libraryDir="libraries" profileDir="profiles" templateDir="templates">
-  <profiles id="org.eclipse.uml2.resources.UML2Profile" description="Basic Profile" uri="platform:/plugin/org.eclipse.uml2.uml.resources/profiles/UML2.profile.uml" exported="false" kind="system"/>
-  <profiles id="org.eclipse.uml2.resources.StandardProfile" description="Complete Profile" uri="platform:/plugin/org.eclipse.uml2.uml.resources/profiles/Standard.profile.uml" exported="false" kind="system"/>
-  <profiles id="com.omondo.uml.core.JavaProfile" description="Java Profile" uri="platform:/plugin/com.omondo.uml.core/profiles/Java.profile.uml" exported="false" kind="system"/>
-  <profiles id="com.omondo.uml2.core.OmondoProfile" description="Omondo Profile" uri="platform:/plugin/com.omondo.uml.core/profiles/Omondo.profile.uml" exported="false" kind="system"/>
-  <profiles id="com.omondo.uml2.core.ArchetypeProfile" description="Archetype Profile" uri="platform:/plugin/com.omondo.uml.core/profiles/Archetype.profile.uml" exported="false" kind="system"/>
-  <profiles id="com.omondo.uml2.core.MetaProfile" description="Meta Profile" uri="platform:/plugin/com.omondo.uml.core/profiles/Meta.profile.uml" exported="false" kind="system"/>
-  <profiles id="com.omondo.32SN.Common" description="Omondo Common Profile" uri="platform:/plugin/com.omondo.uml.core/profiles/Omondo_Common_3.2_SN.profile.uml" exported="false" kind="system"/>
-  <profiles id="com.omondo.32SN.Meta" description="Omondo META Profile" uri="platform:/plugin/com.omondo.uml.core/profiles/Omondo_Meta_3.2_SN.profile.uml" exported="false" kind="system"/>
-  <profiles id="com.omondo.32SN.Persistence" description="Omondo Persistence Profile" uri="platform:/plugin/com.omondo.uml.core/profiles/Omondo_Persistence_3.2_SN.profile.uml" exported="false" kind="system"/>
-  <profiles id="com.omondo.32SN.Presentation" description="Omondo Presentation Profile" uri="platform:/plugin/com.omondo.uml.core/profiles/Omondo_Presentation_3.2_SN.profile.uml" exported="false" kind="system"/>
-  <profiles id="com.omondo.32SN.Process" description="Omondo Process Profile" uri="platform:/plugin/com.omondo.uml.core/profiles/Omondo_Process_3.2_SN.profile.uml" exported="false" kind="system"/>
-  <profiles id="com.omondo.32SN.Service" description="Omondo Service Profile" uri="platform:/plugin/com.omondo.uml.core/profiles/Omondo_Service_3.2_SN.profile.uml" exported="false" kind="system"/>
-  <profiles id="com.omondo.32SN.Webservice" description="Omondo Webservice Profile" uri="platform:/plugin/com.omondo.uml.core/profiles/Omondo_Webservice_3.2_SN.profile.uml" exported="false" kind="system"/>
-  <profiles id="com.omondo.32SN.Xml" description="Omondo XML Profile" uri="platform:/plugin/com.omondo.uml.core/profiles/Omondo_Xml_3.2_SN.profile.uml" exported="false" kind="system"/>
-  <profiles id="com.omondo.uml2.core.DeploymentProfile" description="Deployment Profile" uri="platform:/plugin/com.omondo.uml.core/profiles/Deployment.profile.uml" exported="false" kind="system"/>
-</properties:ProjectDescription>
