+++ simply/src/main/java/it/silma/simply/core/Coefficient.java
