+++ simply/src/main/java/it/silma/simply/core/Variable.java
