+++ simply/src/main/java/it/silma/simply/gui/AboutDialog.java
