+++ simply/src/main/java/it/silma/simply/gui/GraphicDialog.java
