+++ simply/src/main/java/it/silma/simply/gui/SolverFrame.java
