+++ simply/src/main/resources/logo.ico
