+++ simply/src/main/resources/readme.html
